Move common components to common folder
