+++ src/components/yemen/pressure/PressureWhiteHouse.jsx
@@ -11,7 +11,7 @@ import {
 } from 'semantic-ui-react';
 
 // components
-import StepMessage from './../../StepMessage';
+import StepMessage from '../../common/StepMessage';
 
 // messages
 import { whiteHouseMessage } from './messages';
